remove dead code from worker package
