+++ src/server/worker/master.go
@@ -885,85 +885,3 @@ func (a *APIServer) runService(ctx context.Context, logger logs.TaggedLogger) er
 		}
 	})
 }
-
-func (a *APIServer) aggregateProcessStats(stats []*pps.ProcessStats) (*pps.AggregateProcessStats, error) {
-	var downloadTime []float64
-	var processTime []float64
-	var uploadTime []float64
-	var downloadBytes []float64
-	var uploadBytes []float64
-	for _, s := range stats {
-		dt, err := types.DurationFromProto(s.DownloadTime)
-		if err != nil {
-			return nil, err
-		}
-		downloadTime = append(downloadTime, float64(dt))
-		pt, err := types.DurationFromProto(s.ProcessTime)
-		if err != nil {
-			return nil, err
-		}
-		processTime = append(processTime, float64(pt))
-		ut, err := types.DurationFromProto(s.UploadTime)
-		if err != nil {
-			return nil, err
-		}
-		uploadTime = append(uploadTime, float64(ut))
-		downloadBytes = append(downloadBytes, float64(s.DownloadBytes))
-		uploadBytes = append(uploadBytes, float64(s.UploadBytes))
-
-	}
-	dtAgg, err := a.aggregate(downloadTime)
-	if err != nil {
-		return nil, err
-	}
-	ptAgg, err := a.aggregate(processTime)
-	if err != nil {
-		return nil, err
-	}
-	utAgg, err := a.aggregate(uploadTime)
-	if err != nil {
-		return nil, err
-	}
-	dbAgg, err := a.aggregate(downloadBytes)
-	if err != nil {
-		return nil, err
-	}
-	ubAgg, err := a.aggregate(uploadBytes)
-	if err != nil {
-		return nil, err
-	}
-	return &pps.AggregateProcessStats{
-		DownloadTime:  dtAgg,
-		ProcessTime:   ptAgg,
-		UploadTime:    utAgg,
-		DownloadBytes: dbAgg,
-		UploadBytes:   ubAgg,
-	}, nil
-}
-
-func (a *APIServer) aggregate(datums []float64) (*pps.Aggregate, error) {
-	logger := logs.NewMasterLogger(a.pipelineInfo)
-	mean, err := stats.Mean(datums)
-	if err != nil {
-		logger.Logf("error aggregating mean: %v", err)
-	}
-	stddev, err := stats.StandardDeviation(datums)
-	if err != nil {
-		logger.Logf("error aggregating std dev: %v", err)
-	}
-	fifth, err := stats.Percentile(datums, 5)
-	if err != nil {
-		logger.Logf("error aggregating 5th percentile: %v", err)
-	}
-	ninetyFifth, err := stats.Percentile(datums, 95)
-	if err != nil {
-		logger.Logf("error aggregating 95th percentile: %v", err)
-	}
-	return &pps.Aggregate{
-		Count:                 int64(len(datums)),
-		Mean:                  mean,
-		Stddev:                stddev,
-		FifthPercentile:       fifth,
-		NinetyFifthPercentile: ninetyFifth,
-	}, nil
-}
